issue-92: added HrmController
