+++ examples/EmployeesMvc/webapp/Views/Shared/_Layout.cshtml
@@ -52,8 +52,8 @@
                                     HRM
                                 </a>
                                 <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-                                    <a class="dropdown-item" asp-area="" asp-controller="Home" asp-action="Employees">@_appSettings.Translate(languageName, "Employees")</a>
-                                    <a class="dropdown-item" asp-area="" asp-controller="Home" asp-action="Vacations">@_appSettings.Translate(languageName, "Vacations")</a>
+                                    <a class="dropdown-item" asp-area="" asp-controller="Hrm" asp-action="Employees">@_appSettings.Translate(languageName, "Employees")</a>
+                                    <a class="dropdown-item" asp-area="" asp-controller="Hrm" asp-action="Vacations">@_appSettings.Translate(languageName, "Vacations")</a>
                                 </div>
                             </li>
                         </ul>
